# openapi.yml
openapi: 3.0.3
info:
  title: Hyper Fleet Native API
  description: |
    Hyper Fleet Native API provides comprehensive management of virtual machines (sandboxes), 
    file system operations, execution environments, and organizational resources.

    All endpoints except `/api/register` require the `X-API-Key` header for authentication.
  version: 1.0.0
  contact:
    name: Hyper Fleet Support
  license:
    name: Proprietary
servers:
  - url: http://localhost:33944/api
    description: Local development server
  - url: https://api.hyperfleet.io/api
    description: Production server
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: "API key for authentication (format: hf_...)"
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        message:
          type: string
          example: Operation completed successfully
        data:
          type: object
          description: Response data (structure varies by endpoint)
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Invalid request
        details:
          type: string
          example: Additional error details

    # Registration
...
    Organization:
      type: object
      properties:
        id:
          type: string
          example: 65ae1234567890abcdef1234
        name:
          type: string
          example: My Organization
        ownerId:
          type: string
          example: 65ae1234567890abcdef1234
        members:
          type: array
          items:
            type: string
        plan:
          type: string
          example: free
        usage:
          type: integer
          example: 5
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
        updatedAt:
          type: string
          format: date-time

    # API Keys
    GenerateAPIKeyRequest:
      type: object
      required:
        - orgId
        - keyName
      properties:
        orgId:
          type: string
          example: 65ae1234567890abcdef1234
        keyName:
          type: string
          example: CI/CD Key
    GeneratedAPIKeyResponse:
      type: object
      properties:
        plainKey:
          type: string
          description: Only returned once at generation time
          example: hf_1234567890abcdef
        keyId:
          type: string
          example: 65ae1234567890abcdef1234
        keyName:
          type: string
          example: CI/CD Key
        orgId:
          type: string
          example: 65ae1234567890abcdef1234
        createdAt:
          type: string
          format: date-time
        expiresIn:
          type: string
          example: Key expires in 1 year
    APIKeyResponse:
      type: object
      properties:
        id:
          type: string
          example: 65ae1234567890abcdef1234
        orgId:
          type: string
          example: 65ae1234567890abcdef1234
        name:
          type: string
          example: CI/CD Key
        createdBy:
          type: string
          example: 65ae1234567890abcdef1234
        createdAt:
          type: string
          format: date-time
        lastUsedAt:
          type: string
          format: date-time
          nullable: true
        isActive:
          type: boolean
          example: true
        updatedAt:
          type: string
          format: date-time

    # Sandbox
...
    ResizeTerminalRequest:
      type: object
      required:
        - cols
        - rows
      properties:
        cols:
          type: integer
          minimum: 1
          maximum: 500
          example: 100
        rows:
          type: integer
          minimum: 1
          maximum: 200
          example: 40

    # Images
...
paths:
...
  /orgs/me:
    get:
      tags:
        - Organizations
      summary: Get current organization
      description: Get the organization associated with the API key
      operationId: getCurrentOrg
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Organization details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orgs/{orgId}/apikeys:
    get:
      tags:
        - Organizations
      summary: List API keys
      description: Get all API keys for an organization
      operationId: listAPIKeys
      security:
        - ApiKeyAuth: []
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
          example: 65ae1234567890abcdef1234
      responses:
        "200":
          description: List of API keys
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/APIKeyResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      tags:
        - Organizations
      summary: Generate new API key
      description: Create a new API key for the organization. The plain key is only returned once.
      operationId: generateAPIKey
      security:
        - ApiKeyAuth: []
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
          example: 65ae1234567890abcdef1234
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateAPIKeyRequest"
      responses:
        "200":
          description: API key generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedAPIKeyResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /sandboxes/{id}/pty/sessions/{sessionId}/resize:
    post:
      tags:
        - Execution
      summary: Resize terminal
      description: Resize the terminal dimensions for a PTY session
      operationId: resizeTerminal
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: 65ae1234567890abcdef1234
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
          example: session-123456
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResizeTerminalRequest"
      responses:
        "200":
          description: Terminal resized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Session not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...